pyproject.toml
[project]
name = "echopanel"
version = "0.1.0"
description = "EchoPanel v0.1 backend and tooling"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "chromadb>=1.5.0",
    "fastapi==0.115.6",
    "huggingface-hub>=0.34.0,<1.0",
    "mlx-audio>=0.2.9",
    "Pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "pywhispercpp>=1.4.1",
    "qwen-asr>=0.0.6",
    "sentence-transformers>=5.2.2",
    "uvicorn[standard]==0.34.0",
    "vllm>=0.7.3",
    "websockets==14.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.2",
]
asr = [
    "faster-whisper==1.0.3",
    "numpy==1.26.4",
    "requests==2.32.3",
    "pywhispercpp>=1.3.0",
]

diarization = [
    "pyannote.audio==3.3.0",
    "torch==2.5.1",
]
hf_tools = [
    "huggingface_hub>=0.34.0,<1.0",
]

ocr = [
    "paddleocr>=2.7.0",
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "scipy>=1.10.0",
    "scikit-image>=0.20.0",
]

[build-system]
requires = ["setuptools>=69.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["server*"]
exclude = ["macapp*", "docs*", "tests*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]
